--- src/main/java/state/ReadyState.java
@@ -1,16 +1,0 @@
-package state;
-
-class ReadyState implements ProcessState {
-
-    public boolean isError() {
-        return false;
-    }
-
-    public boolean isWarning() {
-        return false;
-    }
-
-    public boolean isReady() {
-        return true;
-    }
-}
