+++ src/main/java/workflow/context/AbstractParameterContext.java
@@ -1,16 +1,21 @@
-package state;
+package workflow.context;
+
+import utils.TypeFactory;
+import workflow.state.ProcessState;
+import workflow.state.ReadyState;
+import workflow.state.StateFactory;
 
 /**
  * An abstract context for configuration objects
  * Will hold details about specific contextual objects
- * Validates and updates its own state
+ * Validates and updates its own workflow
  */
 abstract class AbstractParameterContext extends AbstractContext {
 
     private Object value; //object form of a primative or other type such as range
 
     /**
-     * Instantiate context with a null state
+     * Instantiate context with a null workflow
      */
     AbstractParameterContext() {
         super(StateFactory.INSTANCE.empty());
@@ -20,7 +25,7 @@ abstract class AbstractParameterContext extends AbstractContext {
     /**
      * an abstract function to be defined in concrete Context's
      * checks if the values meet the criteria of the specific context
-     * to be used for updating state
+     * to be used for updating workflow
      * @return boolean
      */
     abstract boolean isValid();
