+++ src/main/java/workflow/state/ProcessState.java
@@ -1,11 +1,11 @@
-package state;
+package workflow.state;
 
 /**
  * An interface to define States that can be seen within the workflow
  * Contains a check for each primary State extension.
  * Only the matching extension should return true on any function
  */
-interface ProcessState {
+public interface ProcessState {
     boolean isError();
     boolean isWarning();
     boolean isReady();
